Added (uint32_t) just in case for older 16bit arduinos.
https://github.com/marcmerlin/Framebuffer_GFX/pull/2/commits/6e73c2fe2f95c00b876aa141587c1916018f221f

+++ Framebuffer_GFX.cpp
@@ -149,7 +149,7 @@ uint16_t Framebuffer_GFX::Color24to16(uint32_t color) {
 }
 
 uint32_t Framebuffer_GFX::CRGBtoint32(CRGB c) {
-  return (c.r<<16) + (c.g<<8) + c.b;
+  return (uint32_t)(c.r<<16) + (uint32_t)(c.g<<8) + (uint32_t)c.b;
 }
